Move LCD screen initialization to occur when knob is pressed.  Make extruder RPM changeable at 0.1 increments.  Display extruder RPM with more precision.  Take away min and max filament diameter.  Update Filament PID settings to best values found so far.  Add splash screen.

+++ Marlin/Marlin.h
@@ -201,8 +201,8 @@ void setPwmFrequency(uint8_t pin, int val);
 #endif //CRITICAL_SECTION_START
 
 extern float extruder_feedrate;
-extern int extruder_rpm; 
-extern int extruder_rpm_set;
+extern float extruder_rpm; 
+extern float extruder_rpm_set;
 extern float puller_feedrate;
 extern float homing_feedrate[];
 extern bool axis_relative_modes[];
